+++ src/main/java/cart/controller/PageController.java
@@ -1,7 +1,7 @@
 package cart.controller;
 
-import cart.controller.dto.MemberResponse;
-import cart.controller.dto.ProductResponse;
+import cart.dto.MemberResponse;
+import cart.dto.ProductResponse;
 import cart.service.MemberService;
 import cart.service.ProductService;
 import java.util.List;
