+++ src/main/java/cart/controller/ProductCartController.java
@@ -1,9 +1,9 @@
 package cart.controller;
 
 import cart.auth.AuthPrincipal;
-import cart.controller.dto.CartRequest;
-import cart.controller.dto.CartResponse;
-import cart.controller.dto.ProductResponse;
+import cart.dto.CartRequest;
+import cart.dto.CartsResponse;
+import cart.dto.ProductResponse;
 import cart.entity.Member;
 import cart.entity.Product;
 import cart.entity.ProductCart;
@@ -12,7 +12,9 @@ import java.net.URI;
 import java.util.List;
 import java.util.stream.Collectors;
 import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -29,12 +31,12 @@ public class ProductCartController {
     }
 
     @GetMapping
-    public ResponseEntity<CartResponse> findMyCart(@AuthPrincipal Member member) {
+    public ResponseEntity<CartsResponse> findMyCart(@AuthPrincipal Member member) {
         List<Product> products = productCartService.findAllMyProductCart(member);
         List<ProductResponse> productResponses = products.stream()
                 .map(ProductResponse::from)
                 .collect(Collectors.toList());
-        return ResponseEntity.ok(new CartResponse(productResponses));
+        return ResponseEntity.ok(new CartsResponse(productResponses));
     }
 
     @PostMapping
@@ -45,4 +47,14 @@ public class ProductCartController {
         ProductCart productCart = productCartService.addCart(cartRequest.getProductId(), member);
         return ResponseEntity.created(URI.create("/carts/" + productCart.getId())).build();
     }
+
+    @DeleteMapping("/{cartId}")
+    public ResponseEntity<Void> deleteMyCart(
+            @PathVariable Long cartId,
+            @AuthPrincipal Member member
+    ) {
+        productCartService.deleteProductInMyCart(cartId, member);
+        return ResponseEntity.noContent().build();
+    }
+
 }
