+++ src/main/java/cart/dto/ProductResponse.java
@@ -1,4 +1,4 @@
-package cart.controller.dto;
+package cart.dto;
 
 import cart.entity.Product;
